requirements.txt
huggingface
openai>=1.0.0
python-dotenv>=0.20.0
numpy
fassio